Дополняет компоненты
